+++ src/poset/caches.go
@@ -119,7 +119,7 @@ func (pec *ParticipantEventsCache) AddPeer(peer *peers.Peer) error {
 	return pec.rim.AddKey(peer.ID)
 }
 
-func (pec *ParticipantEventsCache) participantID(participant string) (uint32, error) {
+func (pec *ParticipantEventsCache) participantID(participant string) (uint64, error) {
 	peer, ok := pec.participants.ByPubKey[participant]
 
 	if !ok {
@@ -200,7 +200,7 @@ func (pec *ParticipantEventsCache) Set(participant string, hash string, index in
 }
 
 // Known returns [participant id] => lastKnownIndex
-func (pec *ParticipantEventsCache) Known() map[uint32]int64 {
+func (pec *ParticipantEventsCache) Known() map[uint64]int64 {
 	return pec.rim.Known()
 }
