+++ src/poset/debug.go
@@ -11,8 +11,8 @@ import (
 func (p *Poset) PrintStat(logger *logrus.Entry) {
 	logger.Warn("****Known events:");
 	for pidID, index := range p.Store.KnownEvents() {
-		logger.Warn("    index=", index, " peer=", p.PeerSet.ById[uint32(pidID)].NetAddr,
-			" pubKeyHex=", p.PeerSet.ById[uint32(pidID)].PubKeyHex)
+		logger.Warn("    index=", index, " peer=", p.PeerSet.ById[pidID].NetAddr,
+			" pubKeyHex=", p.PeerSet.ById[pidID].PubKeyHex)
 	}
 }
